Refactor code to use Layout for passing metaboxes
