+++ packages/story-editor/src/index.js
@@ -21,6 +21,7 @@ import { default as StoryEditor } from './editorApp';
 import PostLock from './components/postLock';
 import StatusCheck from './components/statusCheck';
 import Tooltip from './components/tooltip';
+import { default as InterfaceSkeleton } from './components/layout';
 
 export * from './components/transform';
 export * from './app/config';
@@ -38,5 +39,5 @@ export { GlobalStyle } from './theme';
 export { default as theme } from './theme'; // @todo To be refactored.
 export { GlobalStyle as CropMoveableGlobalStyle } from './components/moveable/cropStyle';
 
-export { PostLock, StatusCheck, Tooltip };
+export { PostLock, StatusCheck, Tooltip, InterfaceSkeleton };
 export default StoryEditor;
